Move modules to correct location
